backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain-groq
python-dotenv
pillow